# Improve performance of the Bazaar VCS backend