--- a/news/5443.bugfix
+++ b/news/5443.bugfix
@@ -1,0 +1,1 @@
+Avoid creating an unnecessary local clone of a Bazaar branch when exporting.