--- a/news/5444.bugfix
+++ b/news/5444.bugfix
@@ -1,0 +1,1 @@
+Use the much faster 'bzr co --lightweight' to obtain a copy of a Bazaar tree.