Improve performance of the Bazaar VCS backend:

* Use lightweight checkouts rather than a full branch clone
* Export directly from the remote branch

This significantly improves performance for Bazaar branches. E.g. for installing bzr itself ("bzr+lp:bzr"):

Performance on my system for 'bzr co --lightweight lp:bzr':

     0.60s user 0.11s system 5% cpu 12.234 total

Performance on my system for 'bzr branch lp:bzr' (current behaviour):

     65.41s user 1.48s system 39% cpu 2:47.91 total

Fixes #5443 
Fixes #5444