@pradyunsg 2018-05-30
Looks like the same as git's --depth parameter and related discussion.