@jelmer 2018-05-30
A distinction between git's --depth parameter and bzr's lightweight checkouts is that history is in bzr's case is not truncated; any access of history (which should be rare) will result in bzr contacting the remote repository to retrieve the missing history (lazily).