@RonnyPfannschmidt 2018-05-31
@pradyunsg basically gits '--depth' is a half-assed feature that breaks half the world, while bzr pretty much did light checkouts correctly