@pradyunsg 2018-06-22
@jelmer Could you link to the issues that this PR should close (using the GitHub auto close 'magic')?

(on mobile)