@cjerdonek 2019-01-06
Here are some high-level comments on this PR. My apologies for not taking a look sooner, but I've never used Bazaar, so I felt like I might not be qualified to review the PR. My comments below are merely high-level.

* This PR changes three different operations for the Bazaar VCS: `export()`, `fetch_new()`, and `update()`. I think they should be proposed and considered separately since the rationale and decision may be different for each, and they may each require different tests, etc.

* For each change there should be at least one test to make sure the code still works with the change. It's possible there aren't yet any functional tests using Bazaar.

* The news files have extension `.bugfix`, but these are performance enhancements and so should be marked `.feature`.

* One of the proposed changes here is to do lightweight checkouts, but for consistency reasons it doesn't seem like we should be doing this for Bazaar if we're not going to do it for Git. And for Git, this idea (shallow clones) was already rejected in the past here: https://github.com/pypa/pip/issues/2432