@jelmer 2019-01-06
As discussed earlier in this pull request, lightweight checkouts in bzr are different from shallow checkouts in Git and the considerations for dropping git shallow clones don't apply here:

A distinction between git's --depth parameter and bzr's lightweight checkouts is that history in bzr's case is not truncated; any access of history (which should be rare) will result in bzr contacting the remote repository to retrieve the missing history (lazily).

I'll see if I can split up this pull request into two separate ones.