@jelmer 2019-01-15
I've split the export part out into a separate proposal: https://github.com/pypa/pip/pull/6139