@pradyunsg 2019-05-23
Closing this since it's bitrotten. Please file a new PR if there's still interest in this.